requirements.txt
streamlit>=1.28.0
sentence-transformers==2.7.0
transformers==4.49.0
tokenizers==0.21.4
timm>=1.0.0
pillow>=7.1.0,<11
faiss-cpu>=1.7.4
openai>=1.3.0
groq>=0.4.0
google-generativeai>=0.3.0
numpy>=1.24.0
torch>=2.0.0
typing_extensions>=4.0.0

